增加截图分享功能

+++ app/src/main/java/cn/vi1zen/zhihudailynew/util/ResUtil.java
@@ -1,11 +1,16 @@
 package cn.vi1zen.zhihudailynew.util;
 
 
+import android.app.Activity;
+import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.Canvas;
 import android.graphics.PixelFormat;
+import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
 import android.os.Environment;
+import android.provider.MediaStore;
+import android.view.View;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -24,15 +29,14 @@ public class ResUtil {
         return App.app.getResources().getString(resId);
     }
 
-    public static String getString(int format, int resId)
-    {
+    public static String getString(int format, int resId) {
         return String.format(App.app.getResources().getString(format), App.app.getResources().getString(resId));
     }
 
-    public static String getString(int format, String content)
-    {
+    public static String getString(int format, String content) {
         return String.format(App.app.getResources().getString(format), content);
     }
+
     public static Bitmap drawableToBitmap(Drawable drawable) {
 
         Bitmap bitmap = Bitmap.createBitmap(
@@ -54,21 +58,63 @@ public class ResUtil {
         return bitmap;
     }
 
-    public static void saveImage(Bitmap bmp) {
+    public static String saveImage(Bitmap bmp,String fileName) {
         File appDir = new File(Constants.STORAGE_DIR);
         if (!appDir.exists()) {
             appDir.mkdir();
         }
-        File file = new File(appDir, Constants.FILE_NAME);
+        File file = new File(appDir, fileName);
+        if(file.exists()){
+            return "";
+        }
         try {
             FileOutputStream fos = new FileOutputStream(file);
             bmp.compress(Bitmap.CompressFormat.JPEG, 100, fos);
             fos.flush();
             fos.close();
+            return file.getAbsolutePath();
         } catch (FileNotFoundException e) {
             e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
         }
+        return "";
+    }
+    public static String insertImageToSystem(Context context, String imagePath) {
+//        MediaStore.Images.Media.query(context.getContentResolver(),imagePath,);
+        String url = "";
+        try {
+            url = MediaStore.Images.Media.insertImage(context.getContentResolver(), imagePath, "测试", "测试图片加入系统图库");
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        }
+        return url;
+    }
+    /**
+     * 截屏
+     * @param activity
+     * @return
+     */
+    public static Bitmap screenShot(Activity activity) {
+        // View是你需要截图的View
+        View view = activity.getWindow().getDecorView();
+        view.setDrawingCacheEnabled(true);
+        view.buildDrawingCache();
+        Bitmap b1 = view.getDrawingCache();
+
+        // 获取状态栏高度
+        Rect frame = new Rect();
+        activity.getWindow().getDecorView().getWindowVisibleDisplayFrame(frame);
+        int statusBarHeight = frame.top;
+
+        // 获取屏幕长和高
+        int width = activity.getWindowManager().getDefaultDisplay().getWidth();
+        int height = activity.getWindowManager().getDefaultDisplay().getHeight();
+        // 去掉标题栏
+        // Bitmap b = Bitmap.createBitmap(b1, 0, 25, 320, 455);
+        Bitmap b = Bitmap.createBitmap(b1, 0, statusBarHeight, width, height
+                - statusBarHeight);
+        view.destroyDrawingCache();
+        return b;
     }
 }
